# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Private Moments API
  version: 0.1.0
  description: API contract for the Mac server used by the iOS app and admin UI.
servers:
  - url: http://127.0.0.1:3210
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/logs:
    get:
      summary: Read recent server logs for the Mac admin UI.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Recent logs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - logs
                properties:
                  logs:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
  /api/v1/admin/posts:
    get:
      summary: List posts for the Mac Admin Posts page.
      parameters:
        - name: deleted
          in: query
          schema:
            type: string
            default: active
            enum:
              - active
              - deleted
              - all
        - name: deviceId
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Admin post list.
          content:
            application/json:
              schema:
                type: object
                required:
                  - posts
                  - nextCursor
                  - searchLimited
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminPost"
                  nextCursor:
                    type:
                      - string
                      - "null"
                  searchLimited:
                    type: boolean
  /api/v1/admin/posts/{postId}:
    get:
      summary: Read one Admin post detail.
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Admin post detail.
          content:
            application/json:
              schema:
                type: object
                required:
                  - post
                properties:
                  post:
                    $ref: "#/components/schemas/AdminPost"
    delete:
      summary: Soft delete one post from Admin.
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Soft-deleted post.
          content:
            application/json:
              schema:
                type: object
                required:
                  - post
                properties:
                  post:
                    anyOf:
                      - $ref: "#/components/schemas/AdminPost"
                      - type: "null"
  /api/v1/admin/devices/{deviceId}/clean-posts/preview:
    get:
      summary: Preview permanent cleanup for posts created by one device.
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cleanup preview.
          content:
            application/json:
              schema:
                type: object
                required:
                  - device
                  - candidateCount
                properties:
                  device:
                    $ref: "#/components/schemas/Device"
                  candidateCount:
                    type: integer
  /api/v1/admin/devices/{deviceId}/clean-posts:
    post:
      summary: Permanently clean posts created by one device.
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - confirmDeviceName
              properties:
                confirmDeviceName:
                  type: string
      responses:
        "200":
          description: Cleanup result.
          content:
            application/json:
              schema:
                type: object
                required:
                  - device
                  - deletedPosts
                  - deletedMediaFiles
                  - failedMediaFiles
                properties:
                  device:
                    $ref: "#/components/schemas/Device"
                  deletedPosts:
                    type: integer
                  deletedMediaFiles:
                    type: integer
                  failedMediaFiles:
                    type: integer
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Device:
      type: object
      required:
        - id
        - name
        - platform
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
        lastSeenAt:
          type:
            - string
            - "null"
          format: date-time
        revokedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Media:
      type: object
      required:
        - id
        - kind
        - status
        - sortOrder
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            - image
            - video
            - audio
        status:
          type: string
          enum:
            - pending
            - uploaded
            - failed
            - deleted
        sortOrder:
          type: integer
        originalPreserved:
          type: boolean
        width:
          type:
            - integer
            - "null"
        height:
          type:
            - integer
            - "null"
        mimeType:
          type:
            - string
            - "null"
        durationSeconds:
          type:
            - number
            - "null"
        transcriptionText:
          type:
            - string
            - "null"
          description: Legacy transcript metadata when present; new iOS does not generate or display it.
        compressedSizeBytes:
          type:
            - integer
            - "null"
        originalSizeBytes:
          type:
            - integer
            - "null"
        checksum:
          type:
            - string
            - "null"
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
        compressedUrl:
          type:
            - string
            - "null"
        originalUrl:
          type:
            - string
            - "null"
        thumbnailUrl:
          type:
            - string
            - "null"
...
    AdminPost:
      type: object
      required:
        - id
        - text
        - isFavorite
        - occurredAt
        - createdAt
        - updatedAt
        - deletedAt
        - serverVersion
        - mediaCount
        - totalMediaCount
        - media
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        isFavorite:
          type: boolean
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
        clientCreatedAt:
          type:
            - string
            - "null"
          format: date-time
        clientUpdatedAt:
          type:
            - string
            - "null"
          format: date-time
        serverVersion:
          type: integer
        createdByDevice:
          anyOf:
            - $ref: "#/components/schemas/Device"
            - type: "null"
        updatedByDevice:
          anyOf:
            - $ref: "#/components/schemas/Device"
            - type: "null"
        mediaCount:
          type: integer
        totalMediaCount:
          type: integer
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
...